refactor: extract requireBoundGuardian to shared auth module
